+++ TabbedPanelMod/history.md
@@ -4,6 +4,8 @@
 
 * added boolean property "isModded" to TabPanel
   * is true if addon have been applied
+* created TPM static class to use it by others addons or scripts
+  * method modifyTab(int tabIndex, String iconName) changes the title to a button with the desired Freeplane icon
 
 ## v0.3